+++ NoskheAPI_Beta/Models/Minimals/Credit.cs
@@ -3,6 +3,6 @@ namespace NoskheAPI_Beta.Models.Minimals
     class Credit
     {
         public string Email { get; set; }
-        public decimal Money { get; set; }
+        public int Money { get; set; }
     }
 }
